+++ ms-review/app/ReviewRepository.py
@@ -1,6 +1,7 @@
 import os
 
-from app.models.Review import ReviewCreate, ReviewOut
+from app.models.ReviewCreate import ReviewCreate
+from app.models.ReviewOut import ReviewOut
 from bson import ObjectId
 from datetime import datetime
 from motor.motor_asyncio import AsyncIOMotorClient
