+++ ms-review/app/models/ReviewCreate.py
@@ -1,0 +1,12 @@
+from datetime import datetime
+from pydantic import BaseModel
+from typing import Optional
+
+
+class ReviewCreate(BaseModel):
+    user_id: Optional[str] = None
+    text: str
+    rating: int
+    image: Optional[str] = None
+    location: Optional[str] = None
+    created_at: Optional[datetime] = datetime.now()
